More warning removal (Some code modifications and cleanup when needed)
--HG--
branch : trunk

+++ src/scripts/outland/auchindoun/shadow_labyrinth/boss_murmur.cpp
@@ -88,7 +88,7 @@ struct boss_murmurAI : public ScriptedAI
     // Sonic Boom instant damage (needs core fix instead of this)
     void SpellHitTarget(Unit *pTarget, const SpellEntry *spell)
     {
-        if (pTarget && pTarget->isAlive() && spell && spell->Id == SPELL_SONIC_BOOM_EFFECT)
+        if (pTarget && pTarget->isAlive() && spell && spell->Id == uint32(SPELL_SONIC_BOOM_EFFECT))
             me->DealDamage(pTarget,(pTarget->GetHealth()*90)/100,NULL,SPELL_DIRECT_DAMAGE,SPELL_SCHOOL_MASK_NATURE,spell);
     }
